requirements.txt
requests
xmltodict
pillow
rocketry
